login/test.py
from argon2 import PasswordHasher

## 비밀번호 확인

if(PasswordHasher().verify('$argon2id$v=19$m=65536,t=3,p=4$vdjtsDtVn4bLRyAHAsCSEw$KKjpF/23s6DFDprcyLTfZrIcIJx58jZt+HYzfErCyO8',
                        'cc')):
    print("true")
