requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
Django==4.1.4
mysqlclient==2.1.1
Pillow==9.4.0
pycparser==2.21
sqlparse==0.4.3
